feat(hypervisor): shared durable filesystem core — mechanical extraction from the GoalRun + OutcomeRoom planes (#73)
Extracts the proven syscall-level mechanics into
hypervisor_daemon_routes/durable_fs.rs, verbatim, with policy staying in the
planes:
- pinned directory descriptors (open_dir_pinned / open_family_dir_pinned) and
  descriptor-relative walks (open_dir_at, mkdir_at, pin_parent) from the
  GoalRun nofollow module (#72 rounds 5-7);
- strict openat(O_NOFOLLOW) slot reads (read_slot_strict, open_file_at,
  bounded read_contained + ReadRefusal);
- the typed atomic-replacement writer (persist_record_durable + PersistFailure,
  #72 rounds 6-8) — moved unchanged, including the promoted-family route and
  dual-write parity;
- the O_TMPFILE descriptor-bound no-clobber commit with unconditional
  family+parent durability barrier and post-barrier inode/byte certification
  (persist_receipt_no_clobber + CommitFailure + certify_target, #72 rounds
  18-21), generalized only by a `family` parameter; the OutcomeRoom plane maps
  CommitFailure variants onto its exact wire codes;
- fd-pinned enumeration (enumerate_pinned) with the NAMED-GAP FIX: readdir EOF
  is distinguished from error via errno (reset before each call, inspected
  after NULL) — an enumeration error propagates typed, never as partial or
  empty truth. Deterministic test with no process-global environment faults
  (errno deliberately polluted before an EOF enumeration; a non-directory fd
  fails typed).

No step-3 objects, routes, fields, or new threat model. GoalRun and OutcomeRoom
keep byte-identical behavior: routes shed 561 lines of local machinery; the
only raw libc call left outside durable_fs in either plane is a test fixture
(mkfifo).

Preservation proof: cargo 196/196; OutcomeRoom verifier 140/140; WorkResult
63/63; operational-depth OK; architecture + pre-next-leg gates green. A journey
(create x2, pause, resume, attach, close, work-result admission) run on the
BASE binary was re-served by the refactored binary with ALL evidence files
byte-identical and zero temp residue, and a pending admission intent sealed by
the base binary was replayed by the refactored binary byte-identically
(reconstruction byte-compare passed; receipt unchanged, room admitted, intent
consumed).
